// src/modules/item/dtos/active-item-dtos/active-item-response-dto.ts
import { Either } from "../../../../helpers/Either";
import DefaultApplicationError from "../../../../presentation/errors/application/default-application-error";

type ItemData = {
	id: string;
	userId: string;
	name: string;
	description: string;
	price: number;
	category: string;
	status: string;
};

export type ActiveItemResponseDto = Either<DefaultApplicationError, ItemData>;